# Faster Triangulated Grid Generation