--- a/src/Meshes.jl
+++ b/src/Meshes.jl
@@ -70,7 +70,7 @@ loadmesh_icosphere_helper(obj_file_name) = EmbeddedDeltaSet2D(
 
 
 # This function was once the gridmeshes.jl file from Decapodes.jl.
-"""    function triangulated_grid(max_x, max_y, dx, dy, point_type, compress=false)
+"""    function triangulated_grid(max_x, max_y, dx, dy, point_type, compress=true)
 
 Triangulate the rectangle [0,max_x] x [0,max_y] by approximately equilateral
 triangles of width `dx` and height `dy`.
@@ -79,56 +79,6 @@ If `compress` is true (default), then enforce that all rows of points are less t
 otherwise, keep `dx` as is.
 """
 function triangulated_grid(max_x, max_y, dx, dy, point_type, compress=true)
-  s = EmbeddedDeltaSet2D{Bool, point_type}()
-
-  # Place equally-spaced points in a max_x by max_y rectangle.
-  coords = point_type == Point3D ? map(x -> point_type(x..., 0), Iterators.product(0:dx:max_x, 0:dy:max_y)) : map(x -> point_type(x...), Iterators.product(0:dx:max_x, 0:dy:max_y))
-  # Perturb every other row right by half a dx.
-  coords[:, 2:2:end] = map(coords[:, 2:2:end]) do row
-    if point_type == Point3D
-      row .+ [dx/2, 0, 0]
-    else
-      row .+ [dx/2, 0]
-    end
-  end
-
-  if compress
-    # The perturbation moved the right-most points past max_x, so compress along x.
-    map!(coords, coords) do coord
-      if point_type == Point3D
-        diagm([max_x/(max_x+dx/2), 1, 1]) * coord
-      else
-        diagm([max_x/(max_x+dx/2), 1]) * coord
-      end
-    end
-  end
-
-  add_vertices!(s, length(coords), point = vec(coords))
-
-  nx = length(0:dx:max_x)
-
-  # Matrix that stores indices of points.
-  idcs = reshape(eachindex(coords), size(coords))
-  # Only grab vertices that will be the bottom-left corner of a subdivided square.
-  idcs = idcs[begin:end-1, begin:end-1]
-
-  # Subdivide every other row along the opposite diagonal.
-  for i in idcs[:, begin+1:2:end]
-    glue_sorted_triangle!(s, i, i+nx, i+nx+1)
-    glue_sorted_triangle!(s, i, i+1, i+nx+1)
-  end
-  for i in idcs[:, begin:2:end]
-    glue_sorted_triangle!(s, i, i+1, i+nx)
-    glue_sorted_triangle!(s, i+1, i+nx, i+nx+1)
-  end
-
-  # Orient and return.
-  s[:edge_orientation]=true
-  orient!(s)
-  s
-end
-
-function new_triangulated_grid(max_x, max_y, dx, dy, point_type, compress=true)
   s = EmbeddedDeltaSet2D{Bool, point_type}()
 
   scale = max_x/(max_x+dx/2)