@lukem12345 2025-01-26
Changing the default edge orientation will change downstream code that relies on flattening vector fields, so after merging we will have to update examples in Decapodes, and check on the Euler page here.